refactor: import json instead of js object
