+++ lib/tests/csrfProtected.test.js
@@ -3,7 +3,7 @@ import { createMocks } from "node-mocks-http";
 import { submitToAPI } from "../helpers";
 import { getCsrfToken } from "next-auth/react";
 import { csrfProtected } from "@lib/middleware/csrfProtected";
-import TestForm from "../../__fixtures__/testData";
+import TestForm from "../../__fixtures__/testData.json";
 jest.mock("axios");
 jest.mock("next-auth/react");
