+++ lib/tests/formBuilder.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import { cleanup, render, screen } from "@testing-library/react";
 import { Form } from "../../components/forms";
 import { getProperty, getFormInitialValues, getRenderedForm } from "../formBuilder";
-import TestForm from "../../__fixtures__/testData";
+import TestForm from "../../__fixtures__/testData.json";
 // GenerateElement from formbuilder is tested in all Component Tests
 
 describe("getProperty() tests", () => {
