+++ lib/tests/helpers.test.js
@@ -7,7 +7,8 @@ import {
 } from "../helpers";
 
 import { logMessage } from "@lib/logger";
-import TestForm from "../../__fixtures__/testData";
+import TestForm from "../../__fixtures__/testData.json";
+
 import { getFormByID } from "../helpers";
 import { getCsrfToken } from "next-auth/react";
